Fix all issues required by code reviewer
